test: 상품 조회 repository 테스트 수정
1. DisplayName 을 구체적으로 하였음.
2. 보다 가독성 좋게 변경하였음.
3. 파일(+클래스) 이름 변경함.
